Live, HTTP tests for an app & device authed user.
Checks that an app_authed user has only specific permissions & fields.
Checks that a device_authed user has access to associations, which in
turn have valid states.

All new tests require that the user has the GEM_API_TOKEN env var set.

The device_auth tests require that a developer has the GEM_EMAIL env var
set. This will create a real user (with a variant'd email address) and
require the tester to complete MFA.

I realize this puts a burden on the tester - but if they do not have
this env var set it will not run these tests. Also, I see these client
tests as a way to ensure the client flow we document is always up to
date with our API.