Remove spec/round completely.
These were unit tests, mocked, or VCRd.

They are mostly out of date and use excessive mocking.
With an API client such as this one, unit tests that rely
so heavily on mocking are essentially useless (IMO).

Useful tests, in this case, are those that document
the use cases for the user and let developers know when
changes in the API need to be reflected in the client.